Add pagination and reorganize component composition for it
